# packages/openapi/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Poveroh API",
        "version": "1.0.0",
        "description": "OpenAPI definition for the Poveroh v1 API"
    },
    "servers": [
        {
            "url": "http://localhost:3001/v1",
            "description": "API Poveroh Localhost v1"
        },
        {
            "url": "http://api.poveroh.local/v1",
            "description": "API Poveroh Localhost proxy v1"
        },
        {
            "url": "http://api.poveroh.com/v1",
            "description": "API Poveroh Production v1"
        }
    ],
    "components": {
        "schemas": {
...
            "StringFilter": {
                "type": "object",
                "properties": {
                    "equals": {
                        "type": "string"
                    },
                    "contains": {
                        "type": "string"
                    }
                }
            },
...
            "FilterOptions": {
                "type": "object",
                "properties": {
                    "skip": {
                        "type": "integer",
                        "minimum": 0,
                        "default": 0
                    },
                    "take": {
                        "type": "integer",
                        "minimum": 0,
                        "exclusiveMinimum": true,
                        "default": 20
                    },
                    "sortBy": {
                        "type": "string",
                        "default": "createdAt"
                    },
                    "sortOrder": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "default": "desc"
                    }
                }
            },
...
            "FinancialAccountTypeEnum": {
                "type": "string",
                "enum": [
                    "ONLINE_BANK",
                    "BANK_ACCOUNT",
                    "CIRCUIT",
                    "DEPOSIT_BANK",
                    "BROKER",
                    "WALLET",
                    "CASH",
                    "CREDIT_CARD",
                    "OTHER"
                ]
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ],
                        "description": "Always false for error responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Error message describing what went wrong"
                    },
                    "error": {
                        "nullable": true,
                        "description": "Optional additional error details or context"
                    }
                },
                "required": [
                    "success",
                    "message"
                ]
            },
...
            "GetFinancialAccountListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FinancialAccountDataResponse"
                        },
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "GetFinancialAccountResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FinancialAccountDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "CreateFinancialAccountRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "number"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                    },
                    "logoIcon": {
                        "type": "string"
                    }
                },
                "required": [
                    "title",
                    "description",
                    "balance",
                    "type",
                    "logoIcon"
                ]
            },
            "CreateFinancialAccountMultipartRequest": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/CreateFinancialAccountRequest"
                    },
                    "file": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary",
                            "description": "Optional file upload"
                        }
                    }
                },
                "required": [
                    "file"
                ]
            },
            "CreateFinancialAccountResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FinancialAccountDataResponse"
                            },
                            {
                                "description": "Response data"
                            }
                        ]
                    }
                },
                "required": [
                    "success"
                ]
            },
            "UpdateFinancialAccountRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "number"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                    },
                    "logoIcon": {
                        "type": "string"
                    }
                }
            },
            "UpdateFinancialAccountResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "DeleteFinancialAccountResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            true
                        ],
                        "description": "Always true for success responses"
                    },
                    "message": {
                        "type": "string",
                        "description": "Optional success message"
                    },
                    "data": {
                        "nullable": true,
                        "description": "Response data"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "FinancialAccountParamsId": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "FinancialAccountFilters": {
                "type": "object",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/FinancialAccountParamsId"
                    },
                    "title": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "description": {
                        "$ref": "#/components/schemas/StringFilter"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                    }
                },
                "additionalProperties": {
                    "anyOf": [
                        {
                            "type": "string"
                        },
                        {
                            "$ref": "#/components/schemas/StringFilter"
                        },
                        {
                            "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                        }
                    ]
                }
            },
...
            "FinancialAccountDataResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "number"
                    },
                    "type": {
                        "$ref": "#/components/schemas/FinancialAccountTypeEnum"
                    },
                    "logoIcon": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "description",
                    "balance",
                    "type",
                    "logoIcon",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/financial-account": {
            "get": {
                "tags": [
                    "Financial Account"
                ],
                "operationId": "getFinancialAccounts",
                "summary": "Get all financial accounts",
                "description": "Retrieve a list of all financial accounts associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/FinancialAccountFilters"
                        },
                        "required": false,
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "$ref": "#/components/schemas/FilterOptions"
                        },
                        "required": false,
                        "name": "options",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of financial accounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetFinancialAccountListResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Financial Account"
                ],
                "operationId": "createFinancialAccount",
                "summary": "Create financial account",
                "description": "Create a new financial account with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Financial account data to create",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFinancialAccountMultipartRequest"
                            },
                            "encoding": {
                                "data": {
                                    "contentType": "application/json"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Financial account created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateFinancialAccountResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Financial Account"
                ],
                "operationId": "deleteFinancialAccounts",
                "summary": "Delete all financial accounts",
                "description": "Delete all financial accounts associated with the user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Financial accounts deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteFinancialAccountResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/financial-accounts/{id}": {
            "get": {
                "tags": [
                    "Financial Account"
                ],
                "operationId": "getFinancialAccountById",
                "summary": "Get financial account by ID",
                "description": "Retrieve a specific financial account by its ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Financial account found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetFinancialAccountResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Financial account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/financial-account/{id}": {
            "patch": {
                "tags": [
                    "Financial Account"
                ],
                "operationId": "updateFinancialAccount",
                "summary": "Update financial account",
                "description": "Update an existing financial account with the provided data",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "description": "Financial account data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFinancialAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Financial account updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateFinancialAccountResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Financial account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Financial Account"
                ],
                "operationId": "deleteFinancialAccount",
                "summary": "Delete financial account",
                "description": "Delete an existing financial account with the provided ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Financial account deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteFinancialAccountResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Financial account not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}